+++ spec/features/staff/users_can_view_programme_level_activity_spec.rb
@@ -1,4 +1,4 @@
-RSpec.feature "Users can view programe level activites" do
+RSpec.feature "Users can view programme level activites" do
   context "when the user belongs to BEIS" do
     let(:user) { create(:beis_user) }
 
@@ -11,7 +11,31 @@ RSpec.feature "Users can view programe level activites" do
         activity: fund_activity)
 
       visit organisation_path(user.organisation)
+
+      expect(page).to have_content I18n.t("page_content.organisation.funds")
+      expect(page).not_to have_content I18n.t("page_content.organisation.programmes")
+
       click_on fund_activity.title
+      click_on programme_activity.title
+
+      page_displays_an_activity(activity_presenter: ActivityPresenter.new(programme_activity))
+    end
+  end
+
+  context "when the user does NOT belong to BEIS" do
+    let(:user) { create(:delivery_partner_user) }
+
+    it "shows the programme level activity" do
+      authenticate!(user: user)
+
+      fund_activity = create(:fund_activity, organisation: user.organisation)
+      programme_activity = create(:programme_activity,
+        organisation: user.organisation,
+        activity: fund_activity)
+
+      visit organisation_path(user.organisation)
+      expect(page).not_to have_content I18n.t("page_content.organisation.funds")
+      expect(page).to have_content I18n.t("page_content.organisation.programmes")
       click_on programme_activity.title
 
       page_displays_an_activity(activity_presenter: ActivityPresenter.new(programme_activity))